// reference/Texas-FrontEnd-TexasInformativeWebsite-release/src/types/resources.ts
export interface PartyPageResourcesProps {
  Home: string;
  party: string;
  "book-your-party": string;
  bothwhitedark: string;
  whitemeat: string;
  darkmeat: string;
  weddingparty: string;
  birthdayparty: string;
  officeparty: string;
  summerbackyardparty: string;
  christmasparty: string;
  watchingtvsportswithyourfriends: string;
  employeeappreciationlunch: string;
  other: string;
  "name-is-required": string;
  "name-must-be-at-most": string;
  "characters-long": string;
  "name-can-only-contain-letters-or-abostrophe": string;
  "phone-number-is-required": string;
  "phone-number-must-be-at-most": string;
  "email-is-required": string;
  "invalid-email-address": string;
  "please-select-location": string;
  "please-select-party-type": string;
  "date-required": string;
  "date-and-time-must-be-in-the-future": string;
  "time-required": string;
  "invalid-time": string;
  "meat-preference-required": string;
  "number-of-chicken-pieces-required": string;
  "number-of-chicken-pieces-must-be-between": string;
  "party-time-must-be-in-the-future": string;
  "successfully-submitted": string;
  "captcha-is-required": string;
  name: string;
  "phone-number": string;
  Email: string;
  Whatkindofpartyareyouplanning: string;
  "preferred-date": string;
  "preferred-time": string;
  "choose-location": string;
  MeatPreference: string;
  Numberofchickenpieces: string;
  Feedback: string;
  "book-now": string;
}